requirements.txt
Django==2.2.0
djangorestframework==3.8.2
psycopg2==2.7.7
black==19.10b0